Move logger test
