+++ .gitignore
@@ -1,3 +1,4 @@
 /vendor
 composer.lock
 phpunit.xml
+.phpunit.result.cache
