+++ src/Logger.php
@@ -12,6 +12,12 @@ class Logger implements StripeLogger
      */
     protected $logger;
 
+    /**
+     * Create a new Logger instance.
+     *
+     * @param  \Psr\Log\LoggerInterface  $logger
+     * @return void
+     */
     public function __construct(LoggerInterface $logger)
     {
         $this->logger = $logger;
